feat(finance): central CostingService + CostCatalogue + AdvanceRequest; integer-cents money
The finance source of truth is now centralized and accuracy-enforced. No
scheduling path computes or persists cost outside the central CostingService.

Money integrity
  * Migrate every money column from FloatField → BigIntegerField integer UGX
    (cost_setting, cost_setting_history, fund_request, fund_request_item,
    monthly_work_plan_budget, admin_budget_line). Data migration floors stray
    fractional values (all live values are whole UGX). The cost engine is
    integer-based throughout; no float money math remains.

CostCatalogue (the active CD Country Cost Catalogue)
  * New CostCatalogue model (country + fy + version + is_active), one active per
    (country, fy). CostSetting rates attach to a catalogue; a data migration
    seeds the active FY catalogue from existing rates. Snapshots stamp
    catalogue_id + catalogue_version onto every budget line.

Central CostingService (apps/budget/costing_service.py — the SINGLE entry point)
  * preview() — itemized cost from the active catalogue with lineItemType,
    catalogue provenance, missingItems + blockers (e.g. 'Group training
    participant meal cost is not set in the active CD Cost Catalogue.').
  * assert_schedulable() — blocks scheduling if a rate is missing, no catalogue,
    or participants absent for training/cluster-meeting. Called by EVERY
    scheduling path before persisting.
  * apply_to_activity() — the canonical budget-line writer: clears + rebuilds
    ActivityScheduleCostLine (now carrying catalogue_id/version/line_item_type/
    currency), re-prices on every schedule/reschedule/partner-self-schedule.

All scheduling paths funnel through it
  * create() now calls assert_schedulable then apply_to_activity (was a duplicate
    _apply_schedule_cost_snapshot). reschedule + partner_schedule both re-price
    (partner_schedule previously never re-priced — a stale-budget gap).
  * New /planning/schedule-cluster-activity endpoint: user chooses Group Training
    vs Cluster Meeting; participants required; cost computed before persist.
    Cluster meeting uses cluster_meeting_cost ONLY (no venue/facilitation; never
    the group-training meal rate) — the critical separation, now enforced.

AdvanceRequest (weekly advance lifecycle, auto-created from budget lines)
  * New AdvanceRequest model + 13-status lifecycle. When an activity is scheduled
    + costed, one AdvanceRequest per budget line is auto-created
    (PENDING_RESPONSIBLE_CONFIRMATION). The responsible user (scheduler) chooses:
    Request Advance / Use Own Funds / Do Not Request. The Accountant can ONLY
    disburse after confirmation (the finance-safety rule). Self-funded advances
    skip disbursement and open a reimbursement path after completion.
  * Endpoints: /fund-requests/advance-queues, /advances/{id}/{confirm-advance,
    self-funded, not-requested, disburse, account, account-approve,
    submit-reimbursement, reimburse}.

Tests: 10 new (test_centralized_costing) prove exact computation for staff
primary/secondary visits, group training, cluster meeting (venue/facilitation
excluded, group-meal-rate NOT used), partner lump sum, missing-rate block,
advance auto-creation, Accountant disbursement gated on confirmation, and the
self-funded → reimbursement path. Backend now 48/48.

+++ edify-api/apps/budget/costing.py
@@ -21,15 +21,15 @@ RateCard = dict  # CostSetting.key -> unitCost
 class CostLine:
     label: str
     key: str
-    unit: float | None  # None = rate missing
+    unit: int | None  # None = rate missing; integer UGX when present
     qty: int
-    amount: float  # 0 when the rate is missing
+    amount: int  # integer UGX (0 when the rate is missing)
     missing: bool
 
 
 @dataclass
 class ActivityCost:
-    amount: float = 0.0
+    amount: int = 0  # integer UGX
     lines: list[CostLine] = field(default_factory=list)
     cost_missing: bool = False
     missing_items: list[str] = field(default_factory=list)
